Separate main code from index.js
